+++ lib/extended_logger.dart
@@ -2,3 +2,4 @@ export 'extended_logger.dart';
 export 'src/___logger.dart';
 export 'src/config.dart';
 export 'src/google_cloud_run_utils.dart';
+export 'src/log_level.dart';