+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+A breaking change.
+
+- Let caller specify how to log locally.
+- Use our own LogLevel enum.
+- Add mappers between LogLevel and other packages log level enums (Logger-package and Talker-package).
+- Don't extend Logger.
+
 ## 0.1.2
 
 - Log in flutter release mode